+++ src/Controller/Component/RememberMeComponent.php
@@ -63,7 +63,7 @@ class RememberMeComponent extends Component
      */
     protected function _validateConfig()
     {
-        if (mb_strlen(Security::salt(), '8bit') < 32) {
+        if (mb_strlen(Security::getSalt(), '8bit') < 32) {
             throw new InvalidArgumentException(
                 __d('CakeDC/Users', 'Invalid app salt, app salt must be at least 256 bits (32 bytes) long')
             );
@@ -77,7 +77,7 @@ class RememberMeComponent extends Component
      */
     protected function _attachEvents()
     {
-        $eventManager = $this->getController()->eventManager();
+        $eventManager = $this->getController()->getEventManager();
         $eventManager->on(UsersAuthComponent::EVENT_AFTER_LOGIN, [], [$this, 'setLoginCookie']);
         $eventManager->on(UsersAuthComponent::EVENT_BEFORE_LOGOUT, [], [$this, 'destroy']);
     }
@@ -105,7 +105,10 @@ class RememberMeComponent extends Component
         if (empty($user)) {
             return;
         }
-        $user['user_agent'] = $this->getController()->request->getHeaderLine('User-Agent');
+        $user['user_agent'] = $this->getController()->getRequest()->getHeaderLine('User-Agent');
+        if (!(bool)$this->getController()->getRequest()->getData(Configure::read('Users.Key.Data.rememberMe'))) {
+             return;
+        }
         $this->Cookie->write($this->_cookieName, $user);
     }
 
@@ -132,10 +135,10 @@ class RememberMeComponent extends Component
     {
         $user = $this->Auth->user();
         if (!empty($user) ||
-            $this->getController()->request->is(['post', 'put']) ||
-            $this->getController()->request->getParam('action') === 'logout' ||
-            $this->getController()->request->session()->check(Configure::read('Users.Key.Session.social')) ||
-            $this->getController()->request->getParam('provider')) {
+            $this->getController()->getRequest()->is(['post', 'put']) ||
+            $this->getController()->getRequest()->getParam('action') === 'logout' ||
+            $this->getController()->getRequest()->getSession()->check(Configure::read('Users.Key.Session.social')) ||
+            $this->getController()->getRequest()->getParam('provider')) {
             return;
         }
 
@@ -149,7 +152,7 @@ class RememberMeComponent extends Component
         if (is_array($event->result)) {
             return $this->getController()->redirect($event->result);
         }
-        $url = $this->getController()->request->getRequestTarget();
+        $url = $this->getController()->getRequest()->getRequestTarget();
 
         return $this->getController()->redirect($url);
     }