+++ tests/TestCase/Controller/Traits/LinkSocialTraitTest.php
@@ -14,9 +14,9 @@ namespace CakeDC\Users\Test\TestCase\Controller\Traits;
 use CakeDC\Users\Controller\Traits\LinkSocialTrait;
 use Cake\Core\Configure;
 use Cake\Event\Event;
+use Cake\Http\Exception\NotFoundException;
 use Cake\Http\ServerRequest;
 use Cake\I18n\Time;
-use Cake\Network\Exception\NotFoundException;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase;
 use League\OAuth2\Client\Provider\Facebook;
@@ -146,7 +146,7 @@ class LinkSocialTraitTest extends BaseTraitTest
                 $this->equalTo([
                     'className' => 'League\OAuth2\Client\Provider\Facebook',
                     'options' => [
-                        'graphApiVersion' => 'v2.5',
+                        'graphApiVersion' => 'v2.8',
                         'redirectUri' => '/auth/facebook',
                         'linkSocialUri' => '/link-social/facebook',
                         'callbackLinkSocialUri' => '/callback-link-social/facebook',
@@ -250,7 +250,7 @@ class LinkSocialTraitTest extends BaseTraitTest
         Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
         Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
 
-        $Table = TableRegistry::get('CakeDC/Users.Users');
+        $Table = TableRegistry::getTableLocator()->get('CakeDC/Users.Users');
 
         $this->Trait = $this->getMockBuilder('CakeDC\Users\Controller\Traits\LinkSocialTrait')
             ->setMethods(['dispatchEvent', 'redirect', 'set', '_createSocialProvider', 'getUsersTable', 'log'])
@@ -335,7 +335,7 @@ class LinkSocialTraitTest extends BaseTraitTest
                 $this->equalTo([
                     'className' => 'League\OAuth2\Client\Provider\Facebook',
                     'options' => [
-                        'graphApiVersion' => 'v2.5',
+                        'graphApiVersion' => 'v2.8',
                         'redirectUri' => '/auth/facebook',
                         'linkSocialUri' => '/link-social/facebook',
                         'callbackLinkSocialUri' => '/callback-link-social/facebook',
@@ -380,8 +380,8 @@ class LinkSocialTraitTest extends BaseTraitTest
     {
         Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
         Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
-        $user = TableRegistry::get('akeDC/Users.Users')->get('00000000-0000-0000-0000-000000000001');
-        $user->errors([
+        $user = TableRegistry::getTableLocator()->get('CakeDC/Users.Users')->get('00000000-0000-0000-0000-000000000001');
+        $user->setErrors([
             'social_accounts' => [
                 '_existsIn' => __d('CakeDC/Users', 'Social account already associated to another user')
             ]
@@ -480,7 +480,7 @@ class LinkSocialTraitTest extends BaseTraitTest
                 $this->equalTo([
                     'className' => 'League\OAuth2\Client\Provider\Facebook',
                     'options' => [
-                        'graphApiVersion' => 'v2.5',
+                        'graphApiVersion' => 'v2.8',
                         'redirectUri' => '/auth/facebook',
                         'linkSocialUri' => '/link-social/facebook',
                         'callbackLinkSocialUri' => '/callback-link-social/facebook',
@@ -507,7 +507,7 @@ class LinkSocialTraitTest extends BaseTraitTest
         Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
         Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
 
-        $Table = TableRegistry::get('CakeDC/Users.Users');
+        $Table = TableRegistry::getTableLocator()->get('CakeDC/Users.Users');
 
         $this->Trait = $this->getMockBuilder('CakeDC\Users\Controller\Traits\LinkSocialTrait')
             ->setMethods(['dispatchEvent', 'redirect', 'set', '_createSocialProvider', 'getUsersTable', 'log'])
@@ -573,7 +573,7 @@ class LinkSocialTraitTest extends BaseTraitTest
                 $this->equalTo([
                     'className' => 'League\OAuth2\Client\Provider\Facebook',
                     'options' => [
-                        'graphApiVersion' => 'v2.5',
+                        'graphApiVersion' => 'v2.8',
                         'redirectUri' => '/auth/facebook',
                         'linkSocialUri' => '/link-social/facebook',
                         'callbackLinkSocialUri' => '/callback-link-social/facebook',
@@ -600,7 +600,7 @@ class LinkSocialTraitTest extends BaseTraitTest
         Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
         Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
 
-        $Table = TableRegistry::get('CakeDC/Users.Users');
+        $Table = TableRegistry::getTableLocator()->get('CakeDC/Users.Users');
 
         $this->Trait = $this->getMockBuilder('CakeDC\Users\Controller\Traits\LinkSocialTrait')
             ->setMethods(['dispatchEvent', 'redirect', 'set', '_createSocialProvider', 'getUsersTable', 'log'])
@@ -664,7 +664,7 @@ class LinkSocialTraitTest extends BaseTraitTest
                 $this->equalTo([
                     'className' => 'League\OAuth2\Client\Provider\Facebook',
                     'options' => [
-                        'graphApiVersion' => 'v2.5',
+                        'graphApiVersion' => 'v2.8',
                         'redirectUri' => '/auth/facebook',
                         'linkSocialUri' => '/link-social/facebook',
                         'callbackLinkSocialUri' => '/callback-link-social/facebook',
@@ -688,7 +688,7 @@ class LinkSocialTraitTest extends BaseTraitTest
         Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
         Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
 
-        $Table = TableRegistry::get('CakeDC/Users.Users');
+        $Table = TableRegistry::getTableLocator()->get('CakeDC/Users.Users');
 
         $this->Trait = $this->getMockBuilder('CakeDC\Users\Controller\Traits\LinkSocialTrait')
             ->setMethods(['dispatchEvent', 'redirect', 'set', '_createSocialProvider', 'getUsersTable', 'log'])
@@ -751,7 +751,7 @@ class LinkSocialTraitTest extends BaseTraitTest
                 $this->equalTo([
                     'className' => 'League\OAuth2\Client\Provider\Facebook',
                     'options' => [
-                        'graphApiVersion' => 'v2.5',
+                        'graphApiVersion' => 'v2.8',
                         'redirectUri' => '/auth/facebook',
                         'linkSocialUri' => '/link-social/facebook',
                         'callbackLinkSocialUri' => '/callback-link-social/facebook',
@@ -775,7 +775,7 @@ class LinkSocialTraitTest extends BaseTraitTest
         Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
         Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
 
-        $Table = TableRegistry::get('CakeDC/Users.Users');
+        $Table = TableRegistry::getTableLocator()->get('CakeDC/Users.Users');
 
         $this->Trait = $this->getMockBuilder('CakeDC\Users\Controller\Traits\LinkSocialTrait')
             ->setMethods(['dispatchEvent', 'redirect', 'set', 'getUsersTable', 'log'])