+++ tests/TestCase/View/Helper/UserHelperTest.php
@@ -131,7 +131,7 @@ class UserHelperTest extends TestCase
      */
     public function testWelcome()
     {
-        $session = $this->getMockBuilder('Cake\Network\Session')
+        $session = $this->getMockBuilder('Cake\Http\Session')
                 ->setMethods(['read'])
                 ->getMock();
         $session->expects($this->at(0))
@@ -145,10 +145,10 @@ class UserHelperTest extends TestCase
             ->will($this->returnValue('david'));
 
         $this->User->request = $this->getMockBuilder('Cake\Network\Request')
-                ->setMethods(['session'])
+                ->setMethods(['getSession'])
                 ->getMock();
         $this->User->request->expects($this->any())
-            ->method('session')
+            ->method('getSession')
             ->will($this->returnValue($session));
 
         $expected = '<span class="welcome">Welcome, <a href="/profile">david</a></span>';
@@ -163,7 +163,7 @@ class UserHelperTest extends TestCase
      */
     public function testWelcomeNotLoggedInUser()
     {
-        $session = $this->getMockBuilder('Cake\Network\Session')
+        $session = $this->getMockBuilder('Cake\Http\Session')
                 ->setMethods(['read'])
                 ->getMock();
         $session->expects($this->at(0))
@@ -172,10 +172,10 @@ class UserHelperTest extends TestCase
             ->will($this->returnValue(null));
 
         $this->User->request = $this->getMockBuilder('Cake\Network\Request')
-                ->setMethods(['session'])
+                ->setMethods(['getSession'])
                 ->getMock();
         $this->User->request->expects($this->any())
-            ->method('session')
+            ->method('getSession')
             ->will($this->returnValue($session));
 
         $result = $this->User->welcome();
@@ -227,10 +227,10 @@ class UserHelperTest extends TestCase
     public function testSocialLoginLink()
     {
         $result = $this->User->socialLogin('facebook');
-        $this->assertEquals('<a href="/auth/facebook" class="btn btn-social btn-facebook "><i class="fa fa-facebook"></i>Sign in with Facebook</a>', $result);
+        $this->assertEquals('<a href="/auth/facebook" class="btn btn-social btn-facebook"><i class="fa fa-facebook"></i>Sign in with Facebook</a>', $result);
 
         $result = $this->User->socialLogin('twitter', ['label' => 'Register with']);
-        $this->assertEquals('<a href="/auth/twitter" class="btn btn-social btn-twitter "><i class="fa fa-twitter"></i>Register with Twitter</a>', $result);
+        $this->assertEquals('<a href="/auth/twitter" class="btn btn-social btn-twitter"><i class="fa fa-twitter"></i>Register with Twitter</a>', $result);
     }
 
     /**
@@ -240,10 +240,10 @@ class UserHelperTest extends TestCase
      */
     public function testSocialLoginTranslation()
     {
-        I18n::locale('es_ES');
+        I18n::setLocale('es_ES');
         $result = $this->User->socialLogin('facebook');
         $this->assertEquals('<a href="/auth/facebook" class="btn btn-social btn-facebook"><i class="fa fa-facebook"></i>Iniciar sesión con Facebook</a>', $result);
-        I18n::locale('en_US');
+        I18n::setLocale('en_US');
     }
 
     /**